requirements.txt
flask==3.1.0
gunicorn==21.2.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-genai==2.1.2
pandas==2.2.3
python-dotenv==1.1.0
werkzeug==3.1.3